Feature/empty dataset (#283)
* update get time functions

* update to fix pylint flake8

* update project dependencies

* fix tests

* update libraries for snyk

* poetry update

* update pytoml

* update test retry time

* add harmony deployment into github actions

* update harmony deployment script

* update pytoml version to develop

* update tests delays

* fix harmony deploy url

* removed debug print statement

* update copy of empty data, and check bounds for empty datasets

* update tests for empty datasets

* update to fix temporal subsetting

* fix ghrsst dataset temporal subsetting

* add try except for getting long name

* optimize mask and scale code

* poetry update

* update changelog

+++ podaac/subsetter/xarray_enhancements.py
@@ -145,8 +145,8 @@ def copy_empty_dataset(dataset: xr.Dataset) -> xr.Dataset:
     # Create a dict object where each key is a variable in the dataset and the value is an
     # array initialized to the fill value for that variable or NaN if there is no fill value
     # attribute for the variable
-    empty_data = {k: np.full(v.shape, dataset.variables[k].attrs.get('_FillValue', np.nan)) for k, v in
-                  dataset.items()}
+
+    empty_data = {k: np.full(v.shape, dataset.variables[k].attrs.get('_FillValue', np.nan), dtype=v.dtype) for k, v in dataset.items()}
 
     # Create a copy of the dataset filled with the empty data. Then select the first index along each
     # dimension and return the result
